# public/swagger_dev.json
{
    "swagger": "2.0",
    "info": {
        "description": "This are APIs server for resume build.",
        "version": "1.0.0",
        "title": "Swagger Resume Builder",
        
        "contact": {
            "email": "ayman.m.elhaggar@gmail.com"
        }
    },
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/resume-templates": {
            "get": {
                "summary": "Get all resume templates",
                "tags": [
                    "Resume Templates"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "Completed Successfully"
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            },
            "post": {
                "summary": "Create new resume template",
                "tags": [
                    "Resume Templates"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "resumeStep",
                        "description": "The resume step to create.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "Required": true,
                                    "default": "Step1"
                                },
                                "title": {
                                    "type": "string",
                                    "Required": true,
                                    "default": "Step 1"
                                },
                                "is_optional": {
                                    "type": "boolean",
                                    "Required": true,
                                    "default": true
                                },
                                "order": {
                                    "type": "number",
                                    "format": "int32",
                                    "Required": true,
                                    "default": 0
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Completed Successfully"
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            }
        },
        "/resume-templates/{id}": {
            "get": {
                "summary": "Get resume template",
                "tags": [
                    "Resume Templates"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the resume step",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "Completed Successfully"
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            },
            "patch": {
                "summary": "Update resume template",
                "tags": [
                    "Resume Templates"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the resume step",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "resumeStep",
                        "description": "The resume step to create.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "default": "Step1"
                                },
                                "title": {
                                    "type": "string",
                                    "default": "Step 1"
                                },
                                "is_optional": {
                                    "type": "boolean",
                                    "default": true
                                },
                                "order": {
                                    "type": "number",
                                    "format": "int32",
                                    "default": 0
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Completed Successfully"
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete resume template",
                "tags": [
                    "Resume Templates"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the resume step",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "Completed Successfully"
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            }
        },
        "/resume-templates/{id}/html": {
            "get": {
                "summary": "Get resume template preview",
                "tags": [
                    "Resume Templates"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the resume step",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "Completed Successfully"
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            }
        },
        "/resume-templates/validate": {
            "post": {
                "summary": "Validate resume template",
                "tags": [
                    "Resume Templates"
                ],
                "parameters": [
                    {
                        "name": "html",
                        "in": "body",
                        "description": "HTML of the resume step template",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "Completed Successfully"
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer Token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "definitions": {
        "Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}